Refactor pattern to add triggers first in the background update
This avoid the pitfalls from events that may be lost in the time
gap if we first populated the `event_stats` table then tried
to add the triggers.

We also get to avoid double-count issues from trying to keep track
where the triggers were added if they happen in the delta itself.