Don't send null values in response to federation profile queries
Don't send the fields `avatar_url` and `displayname` when
they are not defined for the queried user.

Before this change they would be sent and set to null in the JSON response object,
which would violate the OpenAPI definitions
(c.f. https://spec.matrix.org/v1.11/server-server-api/#get_matrixfederationv1queryprofile).