Update `decode_body` to support new lxml behaviour
Updates `decode_body` in the URL previewing code to support the new
behaviour of lxml, where if `etree.fromstring` fails to decode an HTML
page, it now returns an empty HTML Tree instead of `None`.

That "empty" HTML tree contains a single `<body>` tag that contains the
content that it failed to parse (i.e. `�` characters).

Tests now pass on both `lxml` 5.4.0 and 6.0.0.