Don't persist known bad room state from /state_ids
If we are told the state before event E is S via `/state_ids`,
we should refuse to persist E if we fail to persist _all events_
in S, else we are knowingly diverging from that server.

This causes room state to diverge on bad federation connectivity.

Discovered via Chaos testing.