Use @cachedList with `_get_e2e_cross_signing_signatures_for_devices`
The use of `@cachedList` requires having a cached, equivalent "single
item" method that represents the key and cached value. Calls to methods
with @cachedList are then processed and items that are already cached
are removed before the @cachedList-decorated method is called. The
single item method should not be directly used.

The return type needed to be changed to a "Mapping" containing a "Sequence" to satisfy the
requirement that cached results must be immutable (lest you risk
mutating the cache).

`@cachedList` also requires returning an dict where the keys
matches that of the items in the cached input list. Our keys are tuples (and you'll later see why this is a pain).