Rework query to get max stream ID to avoid using a python loop
We now:

1. Group state deltas by `stream_id` and get their count
2. Count (in the DB) until we go over our limit.

We then get the `stream_id` we could naively go up to, as well as the
clamped `stream_id` which would keep us under our limit.

The second query fetches rows up to and including the clamped
`stream_id`.

We also add a unit test that injects multiple state deltas with the same
`stream_id`, which correctly failed when tested against the previous
implementation in this PR.