Validate requests to `/keys/upload` with pydantic
I really wanted to use a `UserID` instead of a `StrictStr` for the fields that contain a user ID. But this became too
cumbersome due to the handler code wanting to directly json-encode the request body. As `UserID` does not subclass
`str`, one would have to rebuild the entire containing object in order to json-encode it. Perhaps a future PR will do this,
as it would allow us to validate UserID's more easily at the edge.