fix rc_room_creation docs
commit 8344c944b1f6ee2f398369fcde7e6f643a1ad6e8 introduced the
option rc_room_creation and the docs are inconsistent,
the default is given as:

```
rc_room_creation:
  per_user:
    per_second: ....
```

while the example shows:

```
rc_room_cration:
  per_second: ...
```

Looking at the source code and the worker example config file,
the latter is correct. So simply remove the stray `per_user:`
and let's pretend it never existed.

The same issue exists with the rc_reports documentation.