Link background process to request trace
This functionality was originally introduced in
https://github.com/element-hq/synapse/pull/18932
but previously worked fine because we were using
the `ContextVarsScopeManager` which had no problem
tracking the tracing scope across logcontext boundaries.
Now that we've switched back to our own custom
`LogContextScopeManager` which tracks the tracing scope
in the logcontext itself, we need to get the active
tracing span before we switch logcontext. (we switch
to the sentinel logcontext because of the `PreserveLoggingContext()`)