Fix indentation of sighup handler calling code
The impact of the bug was that `READY=1` would never be signaled to
systemd. This meant that any `reload` command sent via systemd would
lock the process into a permanent `reloading` state, meaning it could
not be `reload`ed again until the service was restarted.

This could cause operational issues when the processes needed to be
reloaded multiple times.