Prevent `bcrypt` from raising a `ValueError` and log
bcrypt 5.0.0 raises a `ValueError` when a password's length is >72
bytes. Before it would silently truncate instead.

This truncation is fine, as a password of length 72 is sufficiently long
that it will still take millions of years to brute force. The pepper
being cut off as well is also fine - its design to lengthen short
passwords.

We may want to think about transitioning away from bcrypt in future (to
scrypt or argon2id), but for now this returns to existing behaviour
without needing to downgrade `bcrypt`.