Remove cibuildwheel `pp38*` skip selector
`pp38*` used to be enabled, until it went EOL. Now we get the following
warning as we're trying to skip a selector which is not enabled:

```
cibuildwheel: Invalid skip selector: 'pp38*'. This selector matches a group that wasn't enabled. Enable it using the `enable` option or remove this selector. This selector will have no effect.
```

Removing the skip selector silences the warning.