Use `gh` to attach assets to release instead
Stop relying on the `softprops/action-gh-release` action and use the
provided `gh` cli to do so instead.

`softprops/action-gh-release` seemed to have a bug where it would fail
to find the existing, untagged release. It would then create a new
release, and we'd end up with two.

`gh` just works, and is already pre-installed. Let's use that instead.