Bump rpds-py from 0.8.10 to 0.28.0, minimum 0.25.0
On the previous version, we received a segfault when running tests. This
was due to `jsonschema`, which is one of the packages that has rpds-py
as a dependency. I've filed an upstream bug to have them raise the
minimum supported `rpds-py` version:
https://github.com/python-jsonschema/jsonschema/issues/1426. That issue
also includes more details on the segfault, if you're curious.

0.25.0 is the first version to support Python 3.14, which we now
support. Hence making it the minimum.