Better checking of rooms to send
When checking which might have had updates, instead of just checking the
in-memory cache of recent data also check the DB tables. This gives a
more accurate view of how much data there is to send down when either a)
the position is old or b) the server has recently restarted and so has a
smaller cache.