Expire older connections with lots of updates
It's generally faster for clients to reset the sliding sync connection
than to try and send lots of data down on the existing connection.

As a heuristic we arbitrarily choose to reset the connection if a) there
are more than 100 rooms with updates and b) the connection hasn't been
used for at least an hour.