Prevent `lint-newsfile` job activating when fixing dependabot PR branch
Previously, if one pushed a commit to one of @dependabot's PRs (to say,
fix some code that started failing after the upgrade), then the
`lint-newsfile` job would run. It would fail as the PR didn't have a
newsfile. Typically, the developer would then have to write a newsfile.
And during release, said newsfile would not end up in the dependency
bump section, requiring manual fiddling.

This PR changes the conditional to skip the job if the PR was authored
by dependabot, regardless of who authored the commit that triggered the
workflow.