Fix Mastodon URL previews not showing anything useful
Fixes #18444. Inside of UrlPreviewer, we need to combine two
dicts (one from oEmbed, and one from HTML) and in Mastodon's case
they were very different. The one from HTML is basically useless,
due to it being a Javascript application. But the oEmbed one has
the actual post content, yet the information from HTML was preferred.

So I flipped which dictionary overlays which, so keys from oEmbed
is preferred over the extracted HTML ones. This seems to be the
original intention judging by the comment. I also updated to the
newer Python synax for merging dictionaries.